src/requirements.txt
pandas
numpy
pillow
pytesseract
requests
tqdm
lightgbm
sentence-transformers
scikit-learn